backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
